fix: Remove handling of uncaught expections
Previously, one of gotta-catch-em-all's reponsibilities was to
handle uncaught expections, however that did not work as expected
as it also kept fatal errors from halting the system, which is the
intended behavior in some cases. This commit focuses on making
gotta-catch-em-all exclusively for handling unhandled rejections.
Also, tests do not seem to pass on Node 5 or below, as the event
callbacks for handling the rejections do not fire. Bumping the
requirements to Node 6+ since I don't anticipate using this package
in legacy versions. This will be reevalutated upon user request.